Migrate HttpResponseExceptionFilter to HttpResponseExceptionMiddleware
